完成查询商家券批次详细 核销商家券接口实体类
